… cellmlmanip/version.txt