## Description
<!--- Describe your changes in detail -->
added chaching using @lru_cache

## Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here, using the 'fixes #<issue>' syntax. -->
Gives a small performance boost for large models

## Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [X] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

## Documentation checklist
<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
- [x] I have updated all documentation in the code where necessary.
- [x] I have checked spelling in all (new) comments and documentation.
- [x] I have added a note to RELEASE.md if relevant (new feature, breaking change, or notable bug fix).

## Testing
- [x] Testing is done automatically and codecov shows test coverage
- [ ] This cannot be tested automatically <!-- describe how it has been tested -->

I tried with several models on my laptopand it's marginally faster:
| Model   |      Time without      |  Time With |
|----------|:----------------------:|-------------:|
| shannon_wang_puglisi_weber_bers_2004 | 17s | 17s |
| ten_tusscher_model_2006_epi| 9s | 9s |
| mahajan_shiferaw_2008| 14s | 14s
| hund_rudy_2004 | 15s | 15s |
| grandi_pasqualini_bers_2010_ss | 13s | 13s |
| ohara_rudy_2011_endo | 35s | 12s |
| paci_hyttinen_aaltosetala_severi_ventricularVersion| 9s | 9s | 
| ohara_rudy_cipa_v1_2017 | 19s | 7s |
| aslanidi_atrial_model_2009 | 17s | 16s |
| pasek_simurda_christe_2006 | 22s | 20s |